updated rag implementation
