+++ backend/assistant/functions/requirements.txt
@@ -1,12 +1,5 @@
-accelerate==1.2.1
-aiohappyeyeballs==2.4.4
-aiohttp==3.11.11
-aiosignal==1.3.2
 annotated-types==0.7.0
 anyio==4.8.0
-async-timeout==5.0.1
-attrs==24.3.0
-beautifulsoup4==4.12.3
 blinker==1.9.0
 CacheControl==0.14.2
 cachetools==5.5.0
@@ -16,23 +9,17 @@ charset-normalizer==3.4.1
 click==8.1.8
 cloudevents==1.9.0
 cryptography==44.0.0
-dataclasses-json==0.6.7
-Deprecated==1.2.15
 deprecation==2.1.0
-dirtyjson==1.0.8
-distro==1.9.0
 exceptiongroup==1.2.2
 filelock==3.16.1
-filetype==1.2.0
 firebase-admin==6.6.0
 firebase-functions==0.1.2
 Flask==3.1.0
 Flask-Cors==5.0.0
-frozenlist==1.5.0
 fsspec==2024.12.0
 functions-framework==3.8.2
 google-api-core==2.24.0
-google-api-python-client==2.157.0
+google-api-python-client==2.158.0
 google-auth==2.37.0
 google-auth-httplib2==0.2.0
 google-cloud-core==2.4.1
@@ -42,7 +29,6 @@ google-crc32c==1.6.0
 google-events==0.14.0
 google-resumable-media==2.7.2
 googleapis-common-protos==1.66.0
-greenlet==3.1.1
 grpcio==1.69.0
 grpcio-status==1.69.0
 gunicorn==23.0.0
@@ -54,93 +40,55 @@ huggingface-hub==0.27.1
 idna==3.10
 itsdangerous==2.2.0
 Jinja2==3.1.5
-jiter==0.8.2
 joblib==1.4.2
 jsonpatch==1.33
 jsonpointer==3.0.0
 langchain-core==0.3.29
 langchain-huggingface==0.1.2
 langsmith==0.2.10
-llama-cloud==0.1.8
-llama-index==0.12.10
-llama-index-agent-openai==0.4.1
-llama-index-cli==0.4.0
-llama-index-core==0.12.10.post1
-llama-index-embeddings-huggingface==0.4.0
-llama-index-embeddings-openai==0.3.1
-llama-index-indices-managed-llama-cloud==0.6.3
-llama-index-llms-huggingface==0.4.2
-llama-index-llms-huggingface-api==0.3.1
-llama-index-llms-openai==0.3.13
-llama-index-multi-modal-llms-openai==0.4.2
-llama-index-program-openai==0.3.1
-llama-index-question-gen-openai==0.3.0
-llama-index-readers-file==0.4.2
-llama-index-readers-llama-parse==0.4.0
-llama-index-vector-stores-pinecone==0.4.2
-llama-parse==0.5.19
 MarkupSafe==3.0.2
-marshmallow==3.24.2
 mpmath==1.3.0
 msgpack==1.1.0
-multidict==6.1.0
-mypy-extensions==1.0.0
-nest-asyncio==1.6.0
 networkx==3.4.2
-nltk==3.9.1
 numpy==1.26.4
-openai==1.59.5
 orjson==3.10.14
 packaging==24.2
-pandas==2.2.3
 pillow==11.1.0
-pinecone-client==5.0.1
-pinecone-plugin-inference==1.1.0
+pinecone==5.4.2
+pinecone-plugin-inference==3.1.0
 pinecone-plugin-interface==0.0.7
-propcache==0.2.1
 proto-plus==1.25.0
 protobuf==5.29.3
 psutil==6.1.1
 pyasn1==0.6.1
 pyasn1_modules==0.4.1
 pycparser==2.22
-pydantic==2.10.4
+pydantic==2.10.5
 pydantic_core==2.27.2
 PyJWT==2.10.1
 pyparsing==3.2.1
-pypdf==5.1.0
 python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
-pytz==2024.2
 PyYAML==6.0.2
 regex==2024.11.6
 requests==2.32.3
 requests-toolbelt==1.0.0
 rsa==4.9
 safetensors==0.5.2
-scikit-learn==1.6.0
-scipy==1.15.0
+scikit-learn==1.6.1
+scipy==1.15.1
 sentence-transformers==3.3.1
 six==1.17.0
 sniffio==1.3.1
-soupsieve==2.6
-SQLAlchemy==2.0.36
-striprtf==0.0.26
 sympy==1.13.3
 tenacity==9.0.0
-text-generation==0.7.0
 threadpoolctl==3.5.0
-tiktoken==0.8.0
 tokenizers==0.21.0
 torch==2.2.2
 tqdm==4.67.1
-transformers==4.47.1
-typing-inspect==0.9.0
+transformers==4.48.0
 typing_extensions==4.12.2
-tzdata==2024.2
 uritemplate==4.1.1
 urllib3==2.3.0
 watchdog==6.0.0
 Werkzeug==3.1.3
-wrapt==1.17.0
-yarl==1.18.3
